Use vast::builder_callback_ref to build regions
Use vast::builder_callback_ref to build regions instead of passing
regions directly to the builders. We do this using a build_region()
function similar to what VAST has (we don't use the same one that VAST
has though because for some reason that leads to linking errors with
gap).