Use expansion table to visit expansions
Use the expansion table to visit all expansions in the stmt visitor
instead of defining custom a visitor for each visitor method.
If we could somehow group RCU MLIR operation types with their names and
number of parameters then we could further improve this by iterating the
RCU types and then calling the appropriate constructor when we find the
name of the RCU op that the found expansion is an expansion of.