Add member for keeping track of function bodies
Add a member to the kernel visitor for keeping track of function bodies,
so that when the visitor can check when it is visiting a function body
for the first time and reset its internal lock counter.

We can't reset the lock level when visiting the function decl because of
the way vast visitors visit the AST: They visit the decls first, and
then then statements. If vast visitors visited the AST using plain DFS
then this wouldn't be an issue and we could just set the lock level
whenever we visit a function decl. Instead, we must record the body for
checking later.