Remove lock level from kernel visitor
Remove the `m_lock_level` field from the kernel visitor because it's not
necessary for determining the bounds critical sections. This also means
that special visitor for the rcu_read_lock()/unlock() calls is no longer
needed because all it did was annotate the call ops with their lock
levels.